Fixed analytics bug

+++ src/main/java/com/nrifintech/cms/dtos/OrderStatusReport.java
@@ -1,5 +1,8 @@
 package com.nrifintech.cms.dtos;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 import com.nrifintech.cms.types.Status;
 
 import lombok.AllArgsConstructor;
@@ -11,5 +14,5 @@ import lombok.NoArgsConstructor;
 @Data
 public class OrderStatusReport {
     private Status status;
-    private int count;
+    private BigInteger count;
 }
